Extract temporary file logic to a separate function
